+++ src/main.rs
@@ -6,6 +6,7 @@ use std::env;
 use std::error::Error;
 
 mod options;
+mod parser;
 
 const BASE: &str = "http://amzn.com/";
 
@@ -23,35 +24,12 @@ async fn main() -> Result<(), Box<dyn Error>> {
     // convert to doc to parse
     let doc = Document::from(&*html);
 
-    // parse page title
-    let title_raw = doc.find(Attr("id", "productTitle")).next().unwrap().text();
-    let title_clean = title_raw.trim();
+    // parse content
+    let title_clean = parser::parse_title(&doc);
+    let ranks = parser::parse_ranks(&doc);
 
-    println!("{}", title_clean);
-
-    // loop all ranking elements
-    let mut ranks = HashMap::new();
-
-    for node in doc.find(Class("zg_hrsr_item")) {
-        // rank
-        let rank = node.find(Class("zg_hrsr_rank")).next().unwrap().text();
-
-        // title
-        let group = node
-            .find(Class("zg_hrsr_ladder").descendant(Name("a")))
-            .next()
-            .unwrap()
-            .text();
-
-        let clean_group = group
-            .replace("&amp;", "")
-            .replace("\n", "")
-            .replace(";", "");
-        let re = Regex::new(r"\s+").unwrap();
-        let after = re.replace_all(clean_group.as_str(), " ");
-        ranks.insert(rank, after.to_string());
-        println!("{:?}", ranks);
-    }
+    println!("{:?}", title_clean);
+    println!("{:?}", ranks);
 
     Ok(())
 }
